все работает,кроме посылки запросов на /user

+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserService.java
@@ -6,6 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.kata.spring.boot_security.demo.models.User;
 import ru.kata.spring.boot_security.demo.repositories.UserRepository;
 
+import java.security.InvalidParameterException;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,7 +18,7 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-@Transactional
+    @Transactional
     public List<User> showAllUsers() {
         return userRepository.findAll();
     }
@@ -29,23 +30,43 @@ public class UserService {
     }
 
     @Transactional
-    public void save(User user) {
+    public boolean save(User user) {
         User userFromDB = userRepository.findByUsername(user.getUsername());
 
         if (userFromDB != null) {
             System.out.println("Пользователь с таким именем уже есть в базе");
+            return false;
         } else {
             user.setPassword(new BCryptPasswordEncoder().encode(user.getPassword()));
             userRepository.save(user);
+            return true;
         }
     }
 
+
     @Transactional
-    public void update(User updatedUser) {
-        userRepository.save(updatedUser);
+    public boolean update(User user) throws InvalidParameterException {
+        User existingUser = userRepository.findByUsername(user.getUsername());
+        if (existingUser != null && existingUser.getId() != (user.getId())) {
+            return false;
+        }
+        if (user.getPassword().isEmpty()) {
+            user.setPassword(userRepository.findById(user.getId()).get().getPassword());
+        } else {
+            user.setPassword(new BCryptPasswordEncoder().encode(user.getPassword()));
+        }
+        userRepository.save(user);
+        return true;
     }
+
+
     @Transactional
-    public void delete(int id) {
+    public boolean delete(int id) {
+        if (userRepository.findById(id).isEmpty()) {
+            return false;
+        }
         userRepository.deleteById(id);
+        return true;
     }
+
 }
